step3 - Station 도메인 객체 제거 -> StationEntity가 해당 역할도 같이 하도록
