--- src/main/java/subway/station/business/model/Station.java
@@ -1,13 +1,0 @@
-package subway.station.business.model;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class Station {
-
-    private Long id;
-    private String name;
-
-}
