+++ src/main/java/subway/station/repository/StationRepository.java
@@ -1,7 +1,7 @@
 package subway.station.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import subway.station.repository.entity.StationEntity;
+import subway.station.repository.entity.Station;
 
-public interface StationRepository extends JpaRepository<StationEntity, Long> {
+public interface StationRepository extends JpaRepository<Station, Long> {
 }
